Add setting to enable/disable the default lsp keybinds
This PR addresses issue #894 it the simplest way possible. It may be
preferable to implement a more complex solution to disable individual
keybinds or allow the user to override them in some other way.

I decided to submit this as a PR first because I didn't want to try
making larger changes without feedback.

Signed-off-by: Ben Aaron Goldberg <ben@benaaron.dev>